requirements.txt
httpx>=0.27.0
rich>=13.7.0
pydantic>=2.5.0
python-dotenv>=1.0.0